Move bin to separate file
